# Conflicts:
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/ClassFile.java
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/ast/RecordPattern.java
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/ast/SwitchStatement.java
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/lookup/LookupEnvironment.java
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/parser/Parser.java
#	org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/parser/Scanner.java
#	org.eclipse.jdt.core/codeassist/org/eclipse/jdt/internal/codeassist/CompletionEngine.java
#	org.eclipse.jdt.core/codeassist/org/eclipse/jdt/internal/codeassist/complete/CompletionScanner.java
#	org.eclipse.jdt.core/dom/org/eclipse/jdt/core/dom/ASTConverter.java
#	org.eclipse.jdt.core/dom/org/eclipse/jdt/core/dom/ASTMatcher.java
#	org.eclipse.jdt.core/search/org/eclipse/jdt/internal/core/search/BasicSearchEngine.java

+++ org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/util/StackMapAttribute.java
@@ -15,8 +15,8 @@ package org.eclipse.jdt.internal.core.util;
 
 import org.eclipse.jdt.core.util.ClassFormatException;
 import org.eclipse.jdt.core.util.IConstantPool;
-import org.eclipse.jdt.core.util.IStackMapFrame;
 import org.eclipse.jdt.core.util.IStackMapAttribute;
+import org.eclipse.jdt.core.util.IStackMapFrame;
 /**
  * Default implementation of IStackMapAttribute.
  * @see IStackMapAttribute
@@ -56,7 +56,7 @@ public class StackMapAttribute
 		} else {
 			this.frames = NO_FRAMES;
 		}
-		final int byteLength = (int) u4At(classFileBytes, 2, offset);
+		final int byteLength = u4At(classFileBytes, 2, offset);
 
 		if (length != 0) {
 			System.arraycopy(classFileBytes, offset + 6, this.bytes = new byte[byteLength], 0, byteLength);
